Moving and merging migration files to be exposed though package.json.
